renamed ExploreData to _ExploreData
