+++ src/routes/tools/degree-days/index.svelte
@@ -66,7 +66,7 @@
   import { getFeature, reverseGeocode } from "~/helpers/geocode";
 
   // Components
-  import ExploreData from "./ExploreData.svelte";
+  import ExploreData from "./_ExploreData.svelte";
   import {
     Header,
     About,
